python/client/requirements.txt
snet.sdk==0.1.5
